+++ apps/chat/src/utils/server/get-common-page-props.ts
@@ -1,5 +1,4 @@
 import { GetServerSideProps } from 'next';
-import { getToken } from 'next-auth/jwt';
 import { getServerSession } from 'next-auth/next';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 
@@ -162,8 +161,6 @@ export const getCommonPageProps: GetServerSideProps = async ({
   if (isIsolatedView) {
     settings.isolatedModelId = params?.get(ISOLATED_MODEL_QUERY_PARAM) || '';
   }
-
-  console.log(await getToken({ req }));
 
   return {
     props: {
